Fix BN<->Integer conversions on 64-bit systems
I was more of a moron than usual here and didn't take into account that the
limb size of 64-bit systems is 64-bits long. Also, while I'm at it, change the
Word32 types (which are still technically correct) to CInt, so that they float
with the ABI, in case that were to change.

tests/DSA.hs is a good exercise of the BN functions and now passes on 64-bit
systems.

darcs-hash:20080229220410-19166-bf00cfd191291b7a96c6130874440af6dc9b2de0.gz