Turn the Session IO inside out
Previously, IO went through the SocketBIO module, out to C land and back into
Haskell land to block in the RTS. Now we give OpenSSL the file descriptor to
the socket directly and handle the non-blocking return codes, waiting on the
correct events with the RTS calls and retrying the OpenSSL functions.

This is because I couldn't figure out how to do locking correctly with the
previous hair-pin setup.

darcs-hash:20080319032815-19166-06ad788718592736eb48e08dae76b980def99e69.gz